Фикс Если сообщение было удалено, бот пытался его изменить

+++ events/music/trackFinished.js
@@ -1,6 +1,6 @@
 module.exports = {
 	name: "trackFinished",
 	async execute(bot, player) {
-		player.message.delete();
+		if (player.message.deletable) player.message.delete();
 	},
 };
